refactor: drop legacy verb compatibility
