+++ internal/cli/cmdtest/builds_parity_test.go
@@ -391,73 +391,3 @@ func TestParityRelatedCommandsValidationErrors(t *testing.T) {
 
 	runValidationTests(t, tests)
 }
-
-func TestBuildsRemovedGetCommandsPointToCanonicalView(t *testing.T) {
-	t.Setenv("ASC_APP_ID", "")
-
-	tests := []struct {
-		name    string
-		args    []string
-		wantErr string
-	}{
-		{
-			name:    "builds app get",
-			args:    []string{"builds", "app", "get"},
-			wantErr: "Error: `asc builds app get` was removed. Use `asc builds app view` instead.",
-		},
-		{
-			name:    "builds pre-release-version get",
-			args:    []string{"builds", "pre-release-version", "get"},
-			wantErr: "Error: `asc builds pre-release-version get` was removed. Use `asc builds pre-release-version view` instead.",
-		},
-		{
-			name:    "builds beta-app-review-submission get",
-			args:    []string{"builds", "beta-app-review-submission", "get"},
-			wantErr: "Error: `asc builds beta-app-review-submission get` was removed. Use `asc builds beta-app-review-submission view` instead.",
-		},
-		{
-			name:    "builds build-beta-detail get",
-			args:    []string{"builds", "build-beta-detail", "get"},
-			wantErr: "Error: `asc builds build-beta-detail get` was removed. Use `asc builds build-beta-detail view` instead.",
-		},
-		{
-			name:    "builds app-encryption-declaration get",
-			args:    []string{"builds", "app-encryption-declaration", "get"},
-			wantErr: "Error: `asc builds app-encryption-declaration get` was removed. Use `asc builds app-encryption-declaration view` instead.",
-		},
-		{
-			name:    "builds uploads get",
-			args:    []string{"builds", "uploads", "get"},
-			wantErr: "Error: `asc builds uploads get` was removed. Use `asc builds uploads view` instead.",
-		},
-		{
-			name:    "builds uploads files get",
-			args:    []string{"builds", "uploads", "files", "get"},
-			wantErr: "Error: `asc builds uploads files get` was removed. Use `asc builds uploads files view` instead.",
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			root := RootCommand("1.2.3")
-			root.FlagSet.SetOutput(io.Discard)
-
-			stdout, stderr := captureOutput(t, func() {
-				if err := root.Parse(test.args); err != nil {
-					t.Fatalf("parse error: %v", err)
-				}
-				err := root.Run(context.Background())
-				if !errors.Is(err, flag.ErrHelp) {
-					t.Fatalf("expected ErrHelp, got %v", err)
-				}
-			})
-
-			if stdout != "" {
-				t.Fatalf("expected empty stdout, got %q", stdout)
-			}
-			if !strings.Contains(stderr, test.wantErr) {
-				t.Fatalf("expected stderr to contain %q, got %q", test.wantErr, stderr)
-			}
-		})
-	}
-}
