+++ internal/cli/cmdtest/commands_test.go
@@ -2288,18 +2288,6 @@ func TestAgeRatingValidationErrors(t *testing.T) {
 			wantHelp: false,
 		},
 		{
-			name:     "age-rating get removed",
-			args:     []string{"age-rating", "get"},
-			wantErr:  "Error: `asc age-rating get` was removed. Use `asc age-rating view` instead.",
-			wantHelp: true,
-		},
-		{
-			name:     "age-rating set removed",
-			args:     []string{"age-rating", "set", "--id", "AGE_ID"},
-			wantErr:  "Error: `asc age-rating set` was removed. Use `asc age-rating edit` instead.",
-			wantHelp: true,
-		},
-		{
 			name:     "age-rating edit conflicting targets",
 			args:     []string{"age-rating", "edit", "--app-info-id", "INFO_ID", "--version-id", "VERSION_ID"},
 			wantErr:  "only one of --app-info-id or --version-id is allowed",
@@ -2322,12 +2310,6 @@ func TestAgeRatingValidationErrors(t *testing.T) {
 			args:     []string{"age-rating", "edit", "--id", "AGE_ID", "--gambling-simulated", "BAD"},
 			wantErr:  "--gambling-simulated must be one of",
 			wantHelp: false,
-		},
-		{
-			name:     "age-rating get removed with conflicting targets",
-			args:     []string{"age-rating", "get", "--app-info-id", "INFO_ID", "--version-id", "VERSION_ID"},
-			wantErr:  "Error: `asc age-rating get` was removed. Use `asc age-rating view` instead.",
-			wantHelp: true,
 		},
 	}
