+++ internal/cli/testflight/beta_feedback.go
@@ -254,258 +254,6 @@ Examples:
 	}
 }
 
-func DeprecatedBetaFeedbackAliasCommand() *ffcli.Command {
-	fs := flag.NewFlagSet("beta-feedback", flag.ExitOnError)
-
-	cmd := &ffcli.Command{
-		Name:       "beta-feedback",
-		ShortUsage: "asc testflight feedback <subcommand> | asc testflight crashes <subcommand>",
-		ShortHelp:  "Compatibility aliases for older feedback paths.",
-		LongHelp: `Compatibility aliases for older feedback and crash paths.
-
-Prefer:
-  asc testflight feedback ...
-  asc testflight crashes ...`,
-		FlagSet:   fs,
-		UsageFunc: shared.DeprecatedUsageFunc,
-		Subcommands: []*ffcli.Command{
-			deprecatedBetaFeedbackCrashSubmissionsAliasCommand(),
-			deprecatedBetaFeedbackScreenshotSubmissionsAliasCommand(),
-			deprecatedBetaFeedbackCrashLogAliasCommand(),
-		},
-		Exec: func(ctx context.Context, args []string) error {
-			return flag.ErrHelp
-		},
-	}
-
-	return hideTestFlightCommand(cmd)
-}
-
-func deprecatedBetaFeedbackCrashSubmissionsAliasCommand() *ffcli.Command {
-	fs := flag.NewFlagSet("crash-submissions", flag.ExitOnError)
-
-	return &ffcli.Command{
-		Name:       "crash-submissions",
-		ShortUsage: "asc testflight crashes <view|delete> [flags]",
-		ShortHelp:  "Compatibility alias: use `asc testflight crashes ...`.",
-		LongHelp:   `Compatibility alias: use ` + "`asc testflight crashes view`" + ` and ` + "`asc testflight crashes delete`" + `.`,
-		FlagSet:    fs,
-		UsageFunc:  shared.DeprecatedUsageFunc,
-		Subcommands: []*ffcli.Command{
-			deprecatedBetaFeedbackCrashSubmissionsGetAliasCommand(),
-			deprecatedBetaFeedbackCrashSubmissionsDeleteAliasCommand(),
-		},
-		Exec: func(ctx context.Context, args []string) error {
-			return flag.ErrHelp
-		},
-	}
-}
-
-func deprecatedBetaFeedbackCrashSubmissionsGetAliasCommand() *ffcli.Command {
-	fs := flag.NewFlagSet("get", flag.ExitOnError)
-
-	id := fs.String("id", "", "Beta feedback crash submission ID")
-	submissionID := fs.String("submission-id", "", "Crash submission ID")
-	output := shared.BindOutputFlags(fs)
-
-	return &ffcli.Command{
-		Name:       "get",
-		ShortUsage: "asc testflight crashes view --submission-id \"SUBMISSION_ID\"",
-		ShortHelp:  "Compatibility alias: use `asc testflight crashes view`.",
-		LongHelp:   "Compatibility alias: use `asc testflight crashes view --submission-id SUBMISSION_ID`.",
-		FlagSet:    fs,
-		UsageFunc:  shared.DeprecatedUsageFunc,
-		Exec: func(ctx context.Context, args []string) error {
-			idValue, err := resolveLegacyAliasID(strings.TrimSpace(*id), strings.TrimSpace(*submissionID), "--submission-id")
-			if err != nil {
-				fmt.Fprintf(os.Stderr, "Error: %s\n", err.Error())
-				return flag.ErrHelp
-			}
-			if idValue == "" {
-				fmt.Fprintln(os.Stderr, "Error: --submission-id is required")
-				return shared.MissingRequiredUsageError()
-			}
-			return runCrashSubmissionView(ctx, idValue, output)
-		},
-	}
-}
-
-func deprecatedBetaFeedbackCrashSubmissionsDeleteAliasCommand() *ffcli.Command {
-	fs := flag.NewFlagSet("delete", flag.ExitOnError)
-
-	id := fs.String("id", "", "Beta feedback crash submission ID")
-	submissionID := fs.String("submission-id", "", "Crash submission ID")
-	confirm := fs.Bool("confirm", false, "Confirm deletion")
-	output := shared.BindOutputFlags(fs)
-
-	return &ffcli.Command{
-		Name:       "delete",
-		ShortUsage: "asc testflight crashes delete --submission-id \"SUBMISSION_ID\" --confirm",
-		ShortHelp:  "Compatibility alias: use `asc testflight crashes delete`.",
-		LongHelp:   "Compatibility alias: use `asc testflight crashes delete --submission-id SUBMISSION_ID --confirm`.",
-		FlagSet:    fs,
-		UsageFunc:  shared.DeprecatedUsageFunc,
-		Exec: func(ctx context.Context, args []string) error {
-			idValue, err := resolveLegacyAliasID(strings.TrimSpace(*id), strings.TrimSpace(*submissionID), "--submission-id")
-			if err != nil {
-				fmt.Fprintf(os.Stderr, "Error: %s\n", err.Error())
-				return flag.ErrHelp
-			}
-			if idValue == "" {
-				fmt.Fprintln(os.Stderr, "Error: --submission-id is required")
-				return shared.MissingRequiredUsageError()
-			}
-			if !*confirm {
-				fmt.Fprintln(os.Stderr, "Error: --confirm is required")
-				return shared.MissingRequiredUsageError()
-			}
-			return runCrashSubmissionDelete(ctx, idValue, output)
-		},
-	}
-}
-
-func deprecatedBetaFeedbackScreenshotSubmissionsAliasCommand() *ffcli.Command {
-	fs := flag.NewFlagSet("screenshot-submissions", flag.ExitOnError)
-
-	return &ffcli.Command{
-		Name:       "screenshot-submissions",
-		ShortUsage: "asc testflight feedback <view|delete> [flags]",
-		ShortHelp:  "Compatibility alias: use `asc testflight feedback ...`.",
-		LongHelp:   "Compatibility alias: use `asc testflight feedback view` and `asc testflight feedback delete`.",
-		FlagSet:    fs,
-		UsageFunc:  shared.DeprecatedUsageFunc,
-		Subcommands: []*ffcli.Command{
-			deprecatedBetaFeedbackScreenshotSubmissionsGetAliasCommand(),
-			deprecatedBetaFeedbackScreenshotSubmissionsDeleteAliasCommand(),
-		},
-		Exec: func(ctx context.Context, args []string) error {
-			return flag.ErrHelp
-		},
-	}
-}
-
-func deprecatedBetaFeedbackScreenshotSubmissionsGetAliasCommand() *ffcli.Command {
-	fs := flag.NewFlagSet("get", flag.ExitOnError)
-
-	id := fs.String("id", "", "Beta feedback screenshot submission ID")
-	submissionID := fs.String("submission-id", "", "Feedback submission ID")
-	output := shared.BindOutputFlags(fs)
-
-	return &ffcli.Command{
-		Name:       "get",
-		ShortUsage: "asc testflight feedback view --submission-id \"SUBMISSION_ID\"",
-		ShortHelp:  "Compatibility alias: use `asc testflight feedback view`.",
-		LongHelp:   "Compatibility alias: use `asc testflight feedback view --submission-id SUBMISSION_ID`.",
-		FlagSet:    fs,
-		UsageFunc:  shared.DeprecatedUsageFunc,
-		Exec: func(ctx context.Context, args []string) error {
-			idValue, err := resolveLegacyAliasID(strings.TrimSpace(*id), strings.TrimSpace(*submissionID), "--submission-id")
-			if err != nil {
-				fmt.Fprintf(os.Stderr, "Error: %s\n", err.Error())
-				return flag.ErrHelp
-			}
-			if idValue == "" {
-				fmt.Fprintln(os.Stderr, "Error: --submission-id is required")
-				return shared.MissingRequiredUsageError()
-			}
-			return runFeedbackSubmissionView(ctx, idValue, output)
-		},
-	}
-}
-
-func deprecatedBetaFeedbackScreenshotSubmissionsDeleteAliasCommand() *ffcli.Command {
-	fs := flag.NewFlagSet("delete", flag.ExitOnError)
-
-	id := fs.String("id", "", "Beta feedback screenshot submission ID")
-	submissionID := fs.String("submission-id", "", "Feedback submission ID")
-	confirm := fs.Bool("confirm", false, "Confirm deletion")
-	output := shared.BindOutputFlags(fs)
-
-	return &ffcli.Command{
-		Name:       "delete",
-		ShortUsage: "asc testflight feedback delete --submission-id \"SUBMISSION_ID\" --confirm",
-		ShortHelp:  "Compatibility alias: use `asc testflight feedback delete`.",
-		LongHelp:   "Compatibility alias: use `asc testflight feedback delete --submission-id SUBMISSION_ID --confirm`.",
-		FlagSet:    fs,
-		UsageFunc:  shared.DeprecatedUsageFunc,
-		Exec: func(ctx context.Context, args []string) error {
-			idValue, err := resolveLegacyAliasID(strings.TrimSpace(*id), strings.TrimSpace(*submissionID), "--submission-id")
-			if err != nil {
-				fmt.Fprintf(os.Stderr, "Error: %s\n", err.Error())
-				return flag.ErrHelp
-			}
-			if idValue == "" {
-				fmt.Fprintln(os.Stderr, "Error: --submission-id is required")
-				return shared.MissingRequiredUsageError()
-			}
-			if !*confirm {
-				fmt.Fprintln(os.Stderr, "Error: --confirm is required")
-				return shared.MissingRequiredUsageError()
-			}
-			return runFeedbackSubmissionDelete(ctx, idValue, output)
-		},
-	}
-}
-
-func deprecatedBetaFeedbackCrashLogAliasCommand() *ffcli.Command {
-	fs := flag.NewFlagSet("crash-log", flag.ExitOnError)
-
-	return &ffcli.Command{
-		Name:       "crash-log",
-		ShortUsage: "asc testflight crashes log [flags]",
-		ShortHelp:  "Compatibility alias: use `asc testflight crashes log`.",
-		LongHelp:   "Compatibility alias: use `asc testflight crashes log --submission-id SUBMISSION_ID`.",
-		FlagSet:    fs,
-		UsageFunc:  shared.DeprecatedUsageFunc,
-		Subcommands: []*ffcli.Command{
-			deprecatedBetaFeedbackCrashLogGetAliasCommand(),
-		},
-		Exec: func(ctx context.Context, args []string) error {
-			return flag.ErrHelp
-		},
-	}
-}
-
-func deprecatedBetaFeedbackCrashLogGetAliasCommand() *ffcli.Command {
-	fs := flag.NewFlagSet("get", flag.ExitOnError)
-
-	id := fs.String("id", "", "Beta feedback crash submission ID")
-	submissionID := fs.String("submission-id", "", "Crash submission ID")
-	output := shared.BindOutputFlags(fs)
-
-	return &ffcli.Command{
-		Name:       "get",
-		ShortUsage: "asc testflight crashes log --submission-id \"SUBMISSION_ID\"",
-		ShortHelp:  "Compatibility alias: use `asc testflight crashes log`.",
-		LongHelp:   "Compatibility alias: use `asc testflight crashes log --submission-id SUBMISSION_ID`.",
-		FlagSet:    fs,
-		UsageFunc:  shared.DeprecatedUsageFunc,
-		Exec: func(ctx context.Context, args []string) error {
-			idValue, err := resolveLegacyAliasID(strings.TrimSpace(*id), strings.TrimSpace(*submissionID), "--submission-id")
-			if err != nil {
-				fmt.Fprintf(os.Stderr, "Error: %s\n", err.Error())
-				return flag.ErrHelp
-			}
-			if idValue == "" {
-				fmt.Fprintln(os.Stderr, "Error: --submission-id is required")
-				return shared.MissingRequiredUsageError()
-			}
-			return runCrashLogBySubmissionID(ctx, idValue, output)
-		},
-	}
-}
-
-func resolveLegacyAliasID(legacyID, canonicalID, canonicalFlag string) (string, error) {
-	switch {
-	case legacyID != "" && canonicalID != "" && legacyID != canonicalID:
-		return "", fmt.Errorf("%s and --id must match", canonicalFlag)
-	case canonicalID != "":
-		return canonicalID, nil
-	default:
-		return legacyID, nil
-	}
-}
-
 func runFeedbackSubmissionView(ctx context.Context, submissionID string, output shared.OutputFlags) error {
 	client, err := shared.GetASCClient()
 	if err != nil {
